refactor(decompose): split webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-phase6-prompt-file-crash-resilience.spec.ts into folder
