+++ webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-phase6-prompt-file-crash-resilience.spec.ts
@@ -15,454 +15,33 @@
  *   5. close_agent tears the session down + deletes the prompt file.
  */
 
-import { test as base, expect, _electron as electron } from "@playwright/test";
-import type { ElectronApplication, Page } from "@playwright/test";
+import type { ElectronApplication } from "@playwright/test";
 import type { GraphNode } from "@vt/graph-model/graph";
-import { getNodeTitle } from "@vt/graph-model/markdown";
-import { execFileSync } from "node:child_process";
-import { createHash } from "node:crypto";
 import * as fs from "node:fs/promises";
-import { statSync } from "node:fs";
-import * as os from "node:os";
 import * as path from "node:path";
 import {
-  FAKE_AGENT_ENTRYPOINT,
-  WEBAPP_ROOT,
-  getCiElectronFlags,
   mcpCallTool,
   mcpRequest,
-  resolveGraphDaemonNodeBin,
-  resolveTmuxSessionNameForTest,
-  resolveTmuxSessionNamesForTest,
-  robustElectronTeardown,
-  safeStopFileWatching,
-  tmuxCommandArgsForTest,
   waitForMcpServer,
   type ExtendedWindow,
 } from "./electron-smoke-helpers";
-
-// ─── Helpers ─────────────────────────────────────────────────────────────────
-
-function tmuxSessionExists(
-  appSupportPath: string,
-  sessionName: string,
-): boolean {
-  try {
-    execFileSync(
-      "tmux",
-      tmuxCommandArgsForTest(
-        [
-          "has-session",
-          "-t",
-          resolveTmuxSessionNameForTest(sessionName, appSupportPath),
-        ],
-        appSupportPath,
-      ),
-      { stdio: "ignore" },
-    );
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function killTmuxSession(appSupportPath: string, sessionName: string): void {
-  for (const resolvedName of resolveTmuxSessionNamesForTest(
-    sessionName,
-    appSupportPath,
-  )) {
-    try {
-      execFileSync(
-        "tmux",
-        tmuxCommandArgsForTest(
-          ["kill-session", "-t", resolvedName],
-          appSupportPath,
-        ),
-        { stdio: "ignore" },
-      );
-    } catch {
-      // already gone
-    }
-  }
-}
-
-function killProcessGroup(pid: number): void {
-  if (process.platform === "win32") return;
-  try {
-    process.kill(-pid, "SIGKILL");
-  } catch {
-    // The process group may already be gone after the main-process SIGKILL.
-  }
-}
-
-// AGENT_NAMES (in graph-model/settings/types.ts) is the round-robin list the
-// spawn flow picks from; stale sessions of these names from prior runs cause
-// M1-fix5's idempotent rebind to mask the prompt-file spawn under test.
-const AGENT_NAMES: ReadonlyArray<string> = [
-  "Aki",
-  "Ama",
-  "Amit",
-  "Amy",
-  "Anna",
-  "Ari",
-  "Ayu",
-  "Ben",
-  "Bob",
-  "Cho",
-  "Dae",
-  "Dan",
-  "Eli",
-  "Emi",
-  "Eva",
-  "Eve",
-  "Fei",
-  "Gia",
-  "Gus",
-  "Hana",
-  "Ian",
-  "Iris",
-  "Ivan",
-  "Ivy",
-  "Jay",
-  "Jin",
-  "John",
-  "Jose",
-  "Juan",
-  "Jun",
-  "Kai",
-  "Kate",
-  "Leo",
-  "Lou",
-  "Luis",
-  "Mary",
-  "Max",
-  "Meg",
-  "Mei",
-  "Mia",
-  "Nia",
-  "Noa",
-  "Omar",
-  "Otto",
-  "Raj",
-  "Ren",
-  "Rex",
-  "Rio",
-  "Sai",
-  "Sam",
-  "Siti",
-  "Tao",
-  "Tara",
-  "Timi",
-  "Uma",
-  "Vic",
-  "Wei",
-  "Xan",
-  "Yan",
-  "Zoe",
-];
-
-function reapStaleTestTmuxSessions(
-  appSupportPath: string,
-  extraNames: ReadonlyArray<string> = [],
-): void {
-  for (const name of [...AGENT_NAMES, ...extraNames]) {
-    if (tmuxSessionExists(appSupportPath, name)) {
-      killTmuxSession(appSupportPath, name);
-    }
-  }
-}
-
-function tmuxPanePid(
-  appSupportPath: string,
-  sessionName: string,
-): number | null {
-  try {
-    const out = execFileSync(
-      "tmux",
-      tmuxCommandArgsForTest(
-        [
-          "list-panes",
-          "-t",
-          resolveTmuxSessionNameForTest(sessionName, appSupportPath),
-          "-F",
-          "#{pane_pid}",
-        ],
-        appSupportPath,
-      ),
-      { encoding: "utf8" },
-    );
-    const first = out.split("\n").find((line) => line.trim().length > 0);
-    return first ? parseInt(first.trim(), 10) : null;
-  } catch {
-    return null;
-  }
-}
-
-// Voicetree creates a `voicetree-{day}-{month}` subfolder inside the watched
-// directory and uses it as VOICETREE_VAULT_PATH for spawned agents — that's
-// where Phase 6 writes prompt files. Resolve it once after file-watching
-// initializes the project.
-async function resolveVaultWriteFolder(parentDir: string): Promise<string> {
-  const entries = await fs.readdir(parentDir, { withFileTypes: true });
-  const voicetreeDir = entries.find(
-    (e) => e.isDirectory() && /^voicetree(-\d{1,2}-\d{1,2})?$/.test(e.name),
-  );
-  if (!voicetreeDir)
-    throw new Error(
-      `No voicetree-* subfolder under ${parentDir}; file watching may not have initialised yet`,
-    );
-  return path.join(parentDir, voicetreeDir.name);
-}
-
-// Single-line prompt — production normalizes env-var values (collapses whitespace),
-// so the on-disk prompt file will be the single-line form regardless. Keeping the
-// source single-line makes the file-contents assertion direct and avoids depending
-// on normalize behaviour.
-function buildFakeAgentPromptWithSentinel(sentinelTitle: string): string {
-  const script = {
-    actions: [
-      {
-        type: "create_node",
-        title: sentinelTitle,
-        summary: "Phase-6 prompt delivery sentinel",
-      },
-      { type: "delay", ms: 120_000 },
-    ],
-  };
-  return `### FAKE_AGENT_SCRIPT ### ${JSON.stringify(script)} ### END_FAKE_AGENT_SCRIPT ###`;
-}
-
-function graphIncludesTitle(
-  graph: { nodes: Record<string, GraphNode> },
-  title: string,
-): boolean {
-  return Object.values(graph.nodes).some(
-    (node) => getNodeTitle(node) === title,
-  );
-}
-
-async function writeSettings(
-  userDataDir: string,
-  projectRoot: string,
-  agentPrompt: string,
-): Promise<void> {
-  // spawn_agent does NOT accept an inline `agentPrompt`; production resolves the
-  // prompt from `INJECT_ENV_VARS.AGENT_PROMPT` (the same path electron-real-agent
-  // tests use). The headless tmux spawn then writes that value to disk via
-  // applyPromptFileToHeadlessSpawn and exposes it as AGENT_PROMPT_FILE.
-  await fs.writeFile(
-    path.join(userDataDir, "settings.json"),
-    JSON.stringify(
-      {
-        agents: [
-          {
-            name: "Fake Agent",
-            command: `node ${FAKE_AGENT_ENTRYPOINT}`,
-          },
-        ],
-        defaultAgent: "Fake Agent",
-        terminalSpawnPathRelativeToWatchedDirectory: "/",
-        INJECT_ENV_VARS: {
-          AGENT_PROMPT: agentPrompt,
-        },
-      },
-      null,
-      2,
-    ),
-    "utf8",
-  );
-
-  await fs.writeFile(
-    path.join(userDataDir, "projects.json"),
-    JSON.stringify(
-      [
-        {
-          id: "phase6-test-project",
-          path: projectRoot,
-          name: path.basename(projectRoot),
-          type: "folder",
-          lastOpened: Date.now(),
-          voicetreeInitialized: true,
-        },
-      ],
-      null,
-      2,
-    ),
-    "utf8",
-  );
-
-  await fs.writeFile(
-    path.join(userDataDir, "voicetree-config.json"),
-    JSON.stringify(
-      {
-        lastDirectory: projectRoot,
-      },
-      null,
-      2,
-    ),
-    "utf8",
-  );
-}
-
-async function launchElectronApp(
-  userDataDir: string,
-  projectRoot: string,
-): Promise<ElectronApplication> {
-  const app = await electron.launch({
-    args: [
-      ...getCiElectronFlags(),
-      "--remote-debugging-port=0",
-      path.join(WEBAPP_ROOT, "dist-electron/main/index.js"),
-      `--user-data-dir=${userDataDir}`,
-      "--open-folder",
-      projectRoot,
-    ],
-    env: {
-      ...process.env,
-      NODE_ENV: "test",
-      ENABLE_PLAYWRIGHT_DEBUG: "0",
-      VOICETREE_VAULT_PATH: projectRoot, // required for reconcileTmuxHeadlessAgents on startup
-      VOICETREE_PERSIST_STATE: "1",
-      VT_GRAPHD_NODE_BIN: resolveGraphDaemonNodeBin(),
-    },
-    timeout: 60_000,
-  });
-  const proc = app.process();
-  proc?.stdout?.on("data", (c: Buffer) =>
-    console.log(`[MAIN] ${c.toString().trimEnd()}`),
-  );
-  proc?.stderr?.on("data", (c: Buffer) =>
-    console.error(`[MAIN ERR] ${c.toString().trimEnd()}`),
-  );
-  return app;
-}
-
-async function discoverMcpUrl(window: Page): Promise<string> {
-  const port = await window.evaluate(async () => {
-    const api = (window as unknown as ExtendedWindow).electronAPI;
-    if (!api) throw new Error("electronAPI not available");
-    return await (
-      api.main as unknown as { getMcpPort: () => Promise<number> }
-    ).getMcpPort();
-  });
-  return `http://127.0.0.1:${port}/mcp`;
-}
-
-async function bootstrapCallerTerminal(
-  window: Page,
-  parentNodeId: string,
-  callerId: string,
-): Promise<void> {
-  const res = await window.evaluate(
-    async ({ parentNodeId: nodeId, callerId: id }) => {
-      const api = (window as unknown as ExtendedWindow).electronAPI;
-      if (!api?.terminal) throw new Error("electronAPI.terminal not available");
-      return await api.terminal.spawn({
-        type: "Terminal",
-        terminalId: id,
-        attachedToContextNodeId: nodeId,
-        terminalCount: 0,
-        title: "E2E Phase-6 Caller",
-        anchoredToNodeId: { _tag: "None" },
-        shadowNodeDimensions: { width: 600, height: 400 },
-        resizable: true,
-        initialCommand: "sleep 120",
-        executeCommand: true,
-        isPinned: true,
-        isDone: false,
-        lastOutputTime: Date.now(),
-        activityCount: 0,
-        parentTerminalId: null,
-        agentName: id,
-        worktreeName: undefined,
-        isHeadless: false,
-      });
-    },
-    { parentNodeId, callerId },
-  );
-  expect(
-    res.success,
-    `caller terminal spawn failed: ${JSON.stringify(res)}`,
-  ).toBe(true);
-}
-
-// ─── Fixture ─────────────────────────────────────────────────────────────────
-
-type PhaseSixPrompt = {
-  readonly sentinelTitle: string;
-  readonly agentPrompt: string;
-};
-
-type PhaseSixFixtures = {
-  projectRoot: string;
-  phaseSixPrompt: PhaseSixPrompt;
-  userDataDir: string;
-};
-
-const test = base.extend<PhaseSixFixtures>({
-  projectRoot: async ({}, use) => {
-    const root = await fs.mkdtemp(path.join(os.tmpdir(), "vt-phase6-vault-"));
-    const vault = path.join(root, "vault");
-    await fs.mkdir(vault, { recursive: true });
-    await fs.writeFile(
-      path.join(vault, "task.md"),
-      "# Phase 6 prompt-file e2e\n",
-      "utf8",
-    );
-    await use(vault);
-    await fs.rm(root, { recursive: true, force: true });
-  },
-
-  phaseSixPrompt: async ({}, use) => {
-    const promptHash = createHash("sha1")
-      .update(`phase6-${Date.now()}-${Math.random()}`)
-      .digest("hex")
-      .slice(0, 12);
-    const sentinelTitle = `PROMPT_RECEIVED_${promptHash}`;
-    await use({
-      sentinelTitle,
-      agentPrompt: buildFakeAgentPromptWithSentinel(sentinelTitle),
-    });
-  },
-
-  userDataDir: async ({ projectRoot, phaseSixPrompt }, use) => {
-    const dir = await fs.mkdtemp(path.join(os.tmpdir(), "vt-phase6-udata-"));
-    await writeSettings(dir, projectRoot, phaseSixPrompt.agentPrompt);
-    await use(dir);
-    await fs.rm(dir, { recursive: true, force: true });
-  },
-});
-
-async function firstLoadedWindow(
-  app: ElectronApplication,
-  timeout = 30_000,
-): Promise<Page> {
-  const window = await app.firstWindow({ timeout });
-  window.on("console", (msg) => {
-    if (!msg.text().includes("Electron Security Warning")) {
-      console.log(`BROWSER [${msg.type()}]:`, msg.text());
-    }
-  });
-  window.on("pageerror", (err) => console.error("PAGE ERROR:", err.message));
-  await window.waitForLoadState("domcontentloaded");
-  return window;
-}
-
-async function cleanupLiveElectronApp(
-  app: ElectronApplication | null,
-): Promise<void> {
-  if (!app) return;
-  try {
-    await safeStopFileWatching(app);
-  } catch {
-    /* may already be dead */
-  }
-  try {
-    await robustElectronTeardown(app);
-  } catch {
-    /* may already be dead */
-  }
-}
+import {
+  bootstrapCallerTerminal,
+  cleanupLiveElectronApp,
+  discoverMcpUrl,
+  expect,
+  firstLoadedWindow,
+  graphIncludesTitle,
+  killProcessGroup,
+  killTmuxSession,
+  launchElectronApp,
+  promptFileMode,
+  reapStaleTestTmuxSessions,
+  resolveVaultWriteFolder,
+  test,
+  tmuxPanePid,
+  tmuxSessionExists,
+} from "./electron-phase6-prompt-file-crash-resilience/helpers";
 
 // ─── Test ─────────────────────────────────────────────────────────────────────
 
@@ -597,7 +176,7 @@ test.describe("Phase 6 prompt-file + crash resilience (M1-rerun-6)", () => {
           },
         )
         .toBe(true);
-      const fileMode = statSync(promptFile).mode & 0o777;
+      const fileMode = promptFileMode(promptFile);
       expect(fileMode, "prompt file must be mode 0600").toBe(0o600);
       const fileContents = await fs.readFile(promptFile, "utf8");
       expect(fileContents, "prompt file must equal AGENT_PROMPT").toBe(
